Fix generation of test data
Test data was generated by chosing a new seed
for the scenario generation function. This
dig not only generate a few extra cells from
the same type of genes, but also generated
a new gene module.

This new version splits the cells of the
generated data into a test and training data set.